StateStore is a property of collector config now
