--- dev/scconfigs/sc.memory.yaml
@@ -1,3 +1,0 @@
-
-state_store:
-  type: memory
