+++ internal/cmd/config/invoke.go
@@ -5,13 +5,11 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/turbolytics/collector/internal/collector"
 	"github.com/turbolytics/collector/internal/config"
-	"github.com/turbolytics/collector/internal/config/sc"
 	"go.uber.org/zap"
 )
 
 func NewInvokeCmd() *cobra.Command {
 	var configPath string
-	var scConfigPath string
 
 	var invokeCmd = &cobra.Command{
 		Use:   "invoke",
@@ -29,19 +27,9 @@ func NewInvokeCmd() *cobra.Command {
 			if err != nil {
 				panic(err)
 			}
-
-			scConfig, err := sc.NewFromFile(
-				scConfigPath,
-				sc.WithLogger(logger),
-			)
-			if err != nil {
-				panic(err)
-			}
-
 			c, err := collector.New(
 				config,
 				collector.WithLogger(logger),
-				collector.WithStateStorer(scConfig.StateStore.Storer),
 			)
 			if err != nil {
 				panic(err)
@@ -55,9 +43,7 @@ func NewInvokeCmd() *cobra.Command {
 	}
 
 	invokeCmd.Flags().StringVarP(&configPath, "config", "c", "", "Path to config file")
-	invokeCmd.Flags().StringVarP(&scConfigPath, "sc-config", "", "", "Path to signals collector config file")
 	invokeCmd.MarkFlagRequired("config")
-	invokeCmd.MarkFlagRequired("sc-config")
 
 	return invokeCmd
 }
